Create Chapter5-task4
Improved version of previous task, now "Who's your father" is able to print a name of grandfather. Other functionality are the same as in privious task 3. Solved with a creation of dictionary key based on dictionary value (grandfather = family[grandson], so we make a key from a value). Plus used dictionary.get() method to avoid errors.